refactor: npx replacement
Closes #224.
