+++ lib/autofix.js
@@ -3,7 +3,6 @@ const path = require('node:path');
 const chalk = require('chalk');
 const prompts = require('prompts');
 const spawn = require('cross-spawn');
-const getPathKey = require('path-key');
 const FS = require('./fs-wrapper');
 const AppState = require('./state');
 const ErrorMessage = require('./error-message');
@@ -12,6 +11,7 @@ const StyledMessage = require('./styled-message');
 const {startReview} = require('./run-review');
 const ProjectDependencies = require('./project-dependencies');
 const exit = require('../vendor/exit');
+const {backwardsCompatiblePath, pathKey} = require('./npx');
 
 /**
  * @import {App, AutofixRequest} from "./types/app"
@@ -190,7 +190,7 @@ function formatFileContent(options, file, filePath) {
       input: file.source,
       env: hasElmFormatPathFlag
         ? process.env
-        : backwardsCompatibleElmFormatEnv()
+        : {...process.env, [pathKey]: backwardsCompatiblePath()}
     }
   );
 
@@ -235,63 +235,5 @@ ${result.stderr.toString()}`,
     // as that is used as the id of the file.
     path: file.path,
     source: fs.readFileSync(filePath, 'utf8')
-  };
-}
-
-// When `--elm-format-path` was _not_ provided, we used to execute
-// `elm-format` like this:
-//
-// 1. Try `npx elm-format`
-// 2. Try `elm-format`
-//
-// Just starting `npx` takes 200 ms though. Luckily `npx` isn’t even
-// necessary, because the common ways of running `elm-review` are:
-//
-// 1. Install everything globally and run just `elm-review`.
-// 2. Install everything locally and run `npx elm-review`.
-// 3. Use the `--elm-format-path`.
-//
-// That’s also the only supported ways we have for the `elm` binary – there we
-// have never tried to execute `npx elm`.
-//
-// `npx` adds all potential `node_modules/.bin` up to current directory to the
-// beginning of PATH, for example:
-//
-//     ❯ npx node -p 'process.env.PATH.split(require("path").delimiter)'
-//     [
-//       '/Users/you/stuff/node_modules/.bin',
-//       '/Users/you/node_modules/.bin',
-//       '/Users/node_modules/.bin',
-//       '/node_modules/.bin',
-//       '/usr/bin',
-//       'etc'
-//     ]
-//
-// So if a user runs `npx elm-review`, when we later try to spawn just
-// `elm-format`, it’ll be found since when spawning we inherit the same PATH.
-//
-// The `npx elm-format` approach has been removed to avoid those unnessary 200 ms,
-// but to stay backwards compatible we prepend the same paths to the beginning
-// of PATH just like `npx` would (see above). This is needed when:
-//
-// - Executing `elm-review` _without_ `npx`.
-// - And expecting a _local_ `elm-format` to be used. That’s an odd use case,
-//   but was supported due to the `npx` approach.
-//
-// This can be removed in a major version.
-function backwardsCompatibleElmFormatEnv() {
-  const pathKey = getPathKey();
-  return {
-    ...process.env,
-    [pathKey]: [
-      ...process
-        .cwd()
-        .split(path.sep)
-        .map((_, index, parts) =>
-          [...parts.slice(0, index + 1), 'node_modules', '.bin'].join(path.sep)
-        )
-        .reverse(),
-      process.env[pathKey]
-    ].join(path.delimiter)
   };
 }
