+++ lib/elm-binary.js
@@ -3,6 +3,7 @@ const chalk = require('chalk');
 const which = require('which');
 const spawn = require('cross-spawn');
 const ErrorMessage = require('./error-message');
+const {backwardsCompatiblePath} = require('./npx');
 
 /**
  * @import {Options} from "./types/options"
@@ -19,7 +20,7 @@ const ErrorMessage = require('./error-message');
 async function getElmBinary(options) {
   if (options.compiler === undefined) {
     try {
-      return await which('elm');
+      return await which('elm', {path: backwardsCompatiblePath()});
     } catch {
       throw new ErrorMessage.CustomError(
         // prettier-ignore
