+++ tsconfig.no-implicit-any.json
@@ -17,6 +17,7 @@
     "lib/sync-get.js",
     "lib/sync-get-worker.js",
     "lib/error-message.js",
+    "lib/path-helpers.js",
     "lib/flags.js",
     "lib/fs-wrapper.js",
     "lib/hash.js",
@@ -24,11 +25,11 @@
     "lib/load-compiled-app.js",
     "lib/min-version.js",
     "lib/new-rule.js",
+    "lib/npx.js",
     "lib/options.js",
     "lib/os-helpers.js",
     "lib/parse-elm.js",
     "lib/parse-elm-worker.js",
-    "lib/path-helpers.js",
     "lib/project-json-files.js",
     "lib/promisify-port.js",
     "lib/report.js",
